Add files via upload
I've addressed "Implement email templates" issue in this pull request. I have created the dynamic Email- template section for your product. It is very user-friendly, in a single click all the text of the template frame chosen by the user will be selected, and easy to transfer and replace the placeholders.
Please feel free to provide feedback.